ERRAI-979: Provide functionality for CSRF protection
CSRF protection on message bus servlets is enabled by property.
Enabling the property creates a CSRF token on session creation.
The token is then required as a header in every bus POST request.

The token can be written to an HTML page as a JavaScript variable
with a filter, or else the client bus can acquire it from a challenge
from the server message bus.

There is also a filter that protects REST endpoints using the same token.
When an Errai REST caller finds the token in a global JavaScript variable,
it will set this as a header for all REST requests.